{{cookiecutter.project_root}}/requirements/base.txt
toml==0.10.2
typer==0.7.0
alembic==1.9.0
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
Mako==1.2.4
MarkupSafe==2.1.1
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
pycodestyle==2.9.1
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
toml==0.10.2
urllib3==1.26.12
python-dateutil==2.8.2
typer==0.7.0